(PC-17825)[PRO] fix: move useGetBitmap hook
