+++ pro/src/new_components/ImageEditor/ImageEditor.stories.tsx
@@ -1,3 +1,4 @@
+/* istanbul ignore file: not needed */
 import type { Story } from '@storybook/react'
 import React from 'react'
